Moved SequencePair to kcl package
