+++ batchconsumer/batcher/message_batcher_test.go
@@ -7,6 +7,8 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/Clever/amazon-kinesis-client-go/kcl"
 )
 
 type batch [][]byte
@@ -37,7 +39,7 @@ func (m *MockSync) waitForFlush(timeout time.Duration) error {
 	}
 }
 
-var mockSequence = SequencePair{big.NewInt(99999), 12345}
+var mockSequence = kcl.SequencePair{big.NewInt(99999), 12345}
 
 func TestBatchingByCount(t *testing.T) {
 	assert := assert.New(t)
@@ -195,70 +197,24 @@ func TestUpdatingSequence(t *testing.T) {
 	expected := new(big.Int)
 
 	t.Log("After AddMessage (seq=1), smallestSeq = 1")
-	batcher.updateSequenceNumbers(SequencePair{big.NewInt(1), 1234})
+	batcher.updateSequenceNumbers(kcl.SequencePair{big.NewInt(1), 1234})
 	expected.SetInt64(1)
 	seq := batcher.SmallestSequencePair()
 	assert.True(expected.Cmp(seq.Sequence) == 0)
 
 	t.Log("After AddMessage (seq=2), smallestSeq = 1 -- not updated because higher")
-	batcher.updateSequenceNumbers(SequencePair{big.NewInt(2), 1234})
+	batcher.updateSequenceNumbers(kcl.SequencePair{big.NewInt(2), 1234})
 	seq = batcher.SmallestSequencePair()
 	assert.True(expected.Cmp(seq.Sequence) == 0)
 
 	t.Log("After AddMessage (seq=1), smallestSeq = 0")
-	batcher.updateSequenceNumbers(SequencePair{big.NewInt(0), 1234})
+	batcher.updateSequenceNumbers(kcl.SequencePair{big.NewInt(0), 1234})
 	expected.SetInt64(0)
 	seq = batcher.SmallestSequencePair()
 	assert.True(expected.Cmp(seq.Sequence) == 0)
 
 	t.Log("Flushing batch clears smallest sequence pair")
-	assert.NoError(batcher.AddMessage([]byte("cdcd"), SequencePair{big.NewInt(2), 1234}))
+	assert.NoError(batcher.AddMessage([]byte("cdcd"), kcl.SequencePair{big.NewInt(2), 1234}))
 	sync.waitForFlush(time.Minute)
 	assert.Nil(batcher.SmallestSequencePair().Sequence)
-}
-
-func TestSequencePairIsLessThan(t *testing.T) {
-	assert := assert.New(t)
-
-	big10 := big.NewInt(10)
-	big5 := big.NewInt(5)
-
-	tests := []struct {
-		left   SequencePair
-		right  SequencePair
-		isLess bool
-	}{
-		{left: SequencePair{nil, 0}, right: SequencePair{nil, 0}, isLess: false},
-		{left: SequencePair{nil, 0}, right: SequencePair{big10, 0}, isLess: false},
-		{left: SequencePair{big10, 0}, right: SequencePair{nil, 0}, isLess: false},
-
-		{left: SequencePair{big5, 0}, right: SequencePair{big10, 0}, isLess: true},
-		{left: SequencePair{big5, 0}, right: SequencePair{big5, 10}, isLess: true},
-
-		{left: SequencePair{big10, 0}, right: SequencePair{big5, 0}, isLess: false},
-		{left: SequencePair{big5, 10}, right: SequencePair{big5, 0}, isLess: false},
-	}
-
-	for _, test := range tests {
-		left := test.left
-		right := test.right
-		t.Logf(
-			"Is <%s, %d> less than <%s, %d>? %t",
-			left.Sequence.String(), left.SubSequence,
-			right.Sequence.String(), right.SubSequence,
-			test.isLess,
-		)
-
-		assert.Equal(test.isLess, left.IsLessThan(right))
-	}
-}
-
-func TestSequencePairEmpty(t *testing.T) {
-	assert := assert.New(t)
-
-	assert.True(SequencePair{nil, 0}.IsEmpty())
-	assert.True(SequencePair{nil, 10000}.IsEmpty())
-
-	assert.False(SequencePair{big.NewInt(10), 0}.IsEmpty())
-	assert.False(SequencePair{big.NewInt(0), 1000}.IsEmpty())
 }
